Bryt ut filtering till egen komponent (issue #108)
